requirements.txt
openpyxl==3.1.5
pillow==10.4.0
pytesseract==0.3.13
opencv-python==4.10.0.84
numpy==2.1.1
pandas==2.2.2
pywin32==306; platform_system == 'Windows'
regex==2024.9.11
pytest==8.3.3
